+++ ui/app/components/transaction-list-item.js
@@ -8,6 +8,7 @@ const explorerLink = require('../../lib/explorer-link')
 const CopyButton = require('./copyButton')
 const vreme = new (require('vreme'))
 const Tooltip = require('./tooltip')
+const numberToBN = require('number-to-bn')
 
 const TransactionIcon = require('./transaction-list-item-icon')
 const ShiftListItem = require('./shift-list-item')
@@ -39,6 +40,8 @@ TransactionListItem.prototype.render = function () {
     txParams = transaction.msgParams
   }
 
+  const nonce = txParams.nonce ? numberToBN(txParams.nonce).toString(10) : ''
+
   const isClickable = ('hash' in transaction && isLinkable) || isPending
   return (
     h(`.transaction-list-item.flex-row.flex-space-between${isClickable ? '.pointer' : ''}`, {
@@ -67,6 +70,22 @@ TransactionListItem.prototype.render = function () {
         }, [
           h(TransactionIcon, { txParams, transaction, isTx, isMsg }),
         ]),
+      ]),
+
+      h(Tooltip, {
+        title: 'Transaction Number',
+        position: 'bottom',
+      }, [
+        h('span', {
+          style: {
+            display: 'flex',
+            cursor: 'normal',
+            flexDirection: 'column',
+            alignItems: 'center',
+            justifyContent: 'center',
+            padding: '10px',
+          },
+        }, nonce),
       ]),
 
       h('.flex-column', {style: {width: '200px', overflow: 'hidden'}}, [