+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Current Master
 
+- Add Transaction Number (nonce) to transaction list.
+- Label the pending tx icon with a tooltip.
 - Fix bug where website filters would pile up and not deallocate when leaving a site.
 
 ## 3.6.5 2017-5-17